requirements.txt
arrow==1.3.0
asgiref==3.9.1
binaryornot==0.4.4
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
cookiecutter==2.6.0
debugpy==1.8.16
Django==5.2.5
django-browser-reload==1.18.0
django-tailwind==4.2.0
filelock==3.18.0
fsspec==2025.7.0
honcho==2.0.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
pillow==11.3.0
psycopg2-binary==2.9.10
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.4
rich==14.1.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
text-unidecode==1.3
torch==2.8.0
torchvision==0.23.0
triton==3.4.0
types-python-dateutil==2.9.0.20250809
typing_extensions==4.14.1
urllib3==2.5.0
